requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
mfrc522==0.0.7
RPi.GPIO==0.7.1
spidev==3.6
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
requests==2.31.0
python-dateutil==2.8.2 